+++ binders/kafka-binder/spring-cloud-stream-binder-kafka-streams/src/test/java/org/springframework/cloud/stream/binder/kafka/streams/InteractiveQueryServiceMultiStateStoreTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2022 the original author or authors.
+ * Copyright 2022-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,6 +21,7 @@ import java.util.function.Consumer;
 
 import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.streams.KafkaStreams;
+import org.apache.kafka.streams.StoreQueryParameters;
 import org.apache.kafka.streams.errors.UnknownStateStoreException;
 import org.apache.kafka.streams.kstream.KStream;
 import org.apache.kafka.streams.kstream.ValueTransformerWithKey;
@@ -30,6 +31,7 @@ import org.apache.kafka.streams.state.QueryableStoreTypes;
 import org.apache.kafka.streams.state.StoreBuilder;
 import org.apache.kafka.streams.state.Stores;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -48,11 +50,14 @@ import org.springframework.kafka.test.context.EmbeddedKafka;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.awaitility.Awaitility.await;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 /**
  * Tests for the {@link InteractiveQueryService} when dealing with multiple KafkaStreams apps and state stores.
  *
  * @author Chris Bono
+ * @author Soby Chacko
  */
 @EmbeddedKafka(topics = {"input1", "input2"})
 class InteractiveQueryServiceMultiStateStoreTests {
@@ -136,6 +141,33 @@ class InteractiveQueryServiceMultiStateStoreTests {
 					.hasMessage("Error retrieving state store: " + STORE_2_NAME)
 					.hasRootCauseInstanceOf(UnknownStateStoreException.class)
 					.hasRootCauseMessage("Store (" + STORE_2_NAME + ") not available to Streams instance");
+		}
+	}
+
+	@Test
+	@SuppressWarnings("unchecked")
+	void storeQueryParameterCustomizerIsApplied() {
+		try (ConfigurableApplicationContext context = new SpringApplicationBuilder()
+			.sources(StoreQueryParameterCustomizerTestApplication.class)
+			.web(WebApplicationType.NONE)
+			.run("--server.port=0",
+				"--spring.jmx.enabled=false",
+				"--spring.cloud.stream.function.definition=app1",
+				"--spring.cloud.stream.function.bindings.app1-in-0=input1",
+				"--spring.cloud.stream.kafka.streams.binder.functions.app1.application-id=storeQueryParameterCustomizerIsAppliedAppId",
+				"--spring.cloud.stream.kafka.streams.binder.brokers=" + embeddedKafka.getBrokersAsString())
+		) {
+			waitForRunningStreams(context.getBean(KafkaStreamsRegistry.class));
+
+			InteractiveQueryService queryService = context.getBean(InteractiveQueryService.class);
+
+			StoreQueryParametersCustomizer<?> storeQueryParametersCustomizer = context.getBean(StoreQueryParametersCustomizer.class);
+			StoreQueryParameters<?> storeQueryParams = StoreQueryParameters.fromNameAndType(STORE_1_NAME, QueryableStoreTypes.keyValueStore());
+
+			when(storeQueryParametersCustomizer.customize(Mockito.any(StoreQueryParameters.class))).thenReturn(storeQueryParams);
+
+			queryService.getQueryableStore(STORE_1_NAME, QueryableStoreTypes.keyValueStore()).get("someKey");
+			verify(storeQueryParametersCustomizer).customize(Mockito.any(StoreQueryParameters.class));
 		}
 	}
 
@@ -229,6 +261,37 @@ class InteractiveQueryServiceMultiStateStoreTests {
 			return new CleanupConfig(false, true);
 		}
 
+	}
+
+	@EnableAutoConfiguration
+	@Configuration(proxyBeanMethods = false)
+	static class StoreQueryParameterCustomizerTestApplication {
+
+		private static final Logger log = LoggerFactory.getLogger(StoreQueryParameterCustomizerTestApplication.class);
+
+
+		@Bean
+		public StoreBuilder<KeyValueStore<String, String>> store1() {
+			return Stores.keyValueStoreBuilder(
+				Stores.persistentKeyValueStore(STORE_1_NAME), Serdes.String(), Serdes.String());
+		}
+
+		@Bean
+		public Consumer<KStream<String, String>> app1() {
+			return s -> s
+				.transformValues(EchoTransformer::new, STORE_1_NAME)
+				.foreach((k, v) -> log.info("Echo {} -> {} into {}", k, v, STORE_1_NAME));
+		}
+
+		@Bean
+		public CleanupConfig cleanupConfig() {
+			return new CleanupConfig(false, true);
+		}
+
+		@Bean
+		StoreQueryParametersCustomizer<?> storeQueryParametersCustomizer() {
+			return Mockito.mock(StoreQueryParametersCustomizer.class);
+		}
 	}
 
 	static class EchoTransformer implements ValueTransformerWithKey<String, String, String> {
